requirements.txt
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
google-api-core==2.24.2
google-auth==2.38.0
google-cloud-core==2.4.3
google-cloud-storage==3.1.0
google-crc32c==1.7.0
google-resumable-media==2.7.2
googleapis-common-protos==1.69.2
idna==3.10
kestra==0.21.13
proto-plus==1.26.1
protobuf==6.30.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
requests==2.32.3
rsa==4.9
urllib3==2.3.0
